Making SideBar More Dynamic
Implemented a way to have multiple lists displayed on the sidebar depending on which page is being viewed. The back button currently works but if you try to press the same page that you are already on, you won't get that sidebar setting to come back.

To test current state,
	Load the page
	Press Create Form - New sidebar shows
	Press back button - Last sidebar should show